LayoutTests/sputnik/Conformance/07_Lexical_Conventions/7.9_Automatic_Semicolon_Insertion/S7.9_A11_T8-expected.txt

CONSOLE MESSAGE: line 79: SyntaxError: Unexpected token 'else'
S7.9_A11_T8

PASS Expected parsing failure

TEST COMPLETE
